Scale down letterFrequencies for fewer tiles.
Using a pool of 144 letters can lead to a mix of pulled letters that has an odd mix.
For example, since there could always be 3 "V"s in the bag, a short game drawing 36
tiles could draw all 3. Arguably, for a 36 tile game, there should only be one V.

calculateFrequencies() isn't perfect as the new frequencies are calculated by division
and floats get messy. For a 36 tile game, 46 tiles wind up in the bag.